# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/car": {
            "get": {
                "description": "Get List Car",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Car"
                ],
                "summary": "Get List Car",
                "operationId": "get_list_car",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GetCarListBody",
                        "schema": {
                            "$ref": "#/definitions/models.GetListCarResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid Argumant",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Create Car",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Car"
                ],
                "summary": "Create Car",
                "operationId": "create_car",
                "parameters": [
                    {
                        "description": "CreateCarRequestBody",
                        "name": "Car",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateCar"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "GetCarBody",
                        "schema": {
                            "$ref": "#/definitions/models.Car"
                        }
                    },
                    "400": {
                        "description": "Invalid Argumant",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/car/{id}": {
            "get": {
                "description": "Get By ID Car",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Car"
                ],
                "summary": "Get By ID Car",
                "operationId": "get_by_id_car",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GetCarBody",
                        "schema": {
                            "$ref": "#/definitions/models.Car"
                        }
                    },
                    "400": {
                        "description": "Invalid Argumant",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Update Car",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Car"
                ],
                "summary": "Update Car",
                "operationId": "update_car",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "UpdateCarRequestBody",
                        "name": "Car",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateCarSwag"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "UpdateCarBody",
                        "schema": {
                            "$ref": "#/definitions/models.Car"
                        }
                    },
                    "400": {
                        "description": "Invalid Argumant",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Car",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Car"
                ],
                "summary": "Delete Car",
                "operationId": "delete_car",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "DeleteCarBody",
                        "schema": {
                            "$ref": "#/definitions/models.Empty"
                        }
                    },
                    "400": {
                        "description": "Invalid Argumant",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/report/branch-share": {
            "get": {
                "description": "branch share",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "branch share",
                "operationId": "Branch Share",
                "responses": {
                    "201": {
                        "description": "GetListBranchShareResponseBody",
                        "schema": {
                            "$ref": "#/definitions/models.GetListBranchShareResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid Argumant",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/report/debtors": {
            "get": {
                "description": "debtors",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "debtors",
                "operationId": "debtors",
                "responses": {
                    "201": {
                        "description": "GetListDebtorResponseBody",
                        "schema": {
                            "$ref": "#/definitions/models.GetListDebtorResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid Argumant",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/report/investor-share": {
            "get": {
                "description": "investor share",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "investor share",
                "operationId": "Intestor Share",
                "responses": {
                    "201": {
                        "description": "GetListInvestorShareResponseBody",
                        "schema": {
                            "$ref": "#/definitions/models.GetListInvestorShareResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid Argumant",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.BranchShare": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        },
        "models.Car": {
            "type": "object",
            "properties": {
                "branch_id": {
                    "type": "string"
                },
                "crated_at": {
                    "type": "string"
                },
                "daily_limit": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "investor_id": {
                    "type": "string"
                },
                "investor_percentage": {
                    "type": "number"
                },
                "km": {
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                },
                "over_limit": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "state_number": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.CreateCar": {
            "type": "object",
            "properties": {
                "branch_id": {
                    "type": "string"
                },
                "daily_limit": {
                    "type": "integer"
                },
                "investor_id": {
                    "type": "string"
                },
                "investor_percentage": {
                    "type": "number"
                },
                "km": {
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                },
                "over_limit": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "state_number": {
                    "type": "string"
                }
            }
        },
...
        "models.Debtor": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        },
        "models.Empty": {
            "type": "object"
        },
...
        "models.GetListBranchShareResponse": {
            "type": "object",
            "properties": {
                "branchs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BranchShare"
                    }
                },
                "count": {
                    "type": "integer"
                }
            }
        },
        "models.GetListCarResponse": {
            "type": "object",
            "properties": {
                "cars": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Car"
                    }
                },
                "count": {
                    "type": "integer"
                }
            }
        },
...
        "models.GetListDebtorResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "debtors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Debtor"
                    }
                }
            }
        },
...
        "models.GetListInvestorShareResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "investors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.InvestorShare"
                    }
                }
            }
        },
...
        "models.InvestorShare": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        },
...
        "models.UpdateCarSwag": {
            "type": "object",
            "properties": {
                "branch_id": {
                    "type": "string"
                },
                "daily_limit": {
                    "type": "integer"
                },
                "investor_id": {
                    "type": "string"
                },
                "investor_percentage": {
                    "type": "number"
                },
                "km": {
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                },
                "over_limit": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "state_number": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}